* Convert Flow Sensor reading to float with unit as litre.

+++ main_code/Nano_main/Nano_UART.h
@@ -41,7 +41,7 @@ void sTempSendToMaster(uint16_t*);
 /**
 Send flow data to master
 **/
-void sFlowSendToMaster(uint16_t*);
+void sFlowSendToMaster(float*);
 // ------ Public variable -------------------------------------
 
 #endif // __NANO_UART_H
